contribution from Michael Peters
In the event that a document had no style block the code ended up croaking rather than just returning back the original html.

The desired behaviour here clearly is not a croak, merging this in.

Original request from CPAN RT here:
https://rt.cpan.org/Public/Bug/Display.html?id=58957

This is not a patch application - I modified to avoid multiple return statements for readability


git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@2638 49aadcc1-c794-da11-82e1-00902723a867