Clever contribution by Michael Peters (wonko) from CPAN RT.
In the previous versions of the library we consciously ignored ordering of css rulesets - mostly because the tools we had
couldn't support the ordering.

Contributor went back to CSS::Tiny and asked the author to change his library around. The library can now provide a
handle through which it can be coupled together with an ordered hash (in this case Tie::IxHash).

This is the initial merge - it's very simple.

I need to get the test facility working after which I can test this code and ensure that it works as expected.

This is unlikely to provide a benefit to MailerMailer's processing as we have fairly explicit rulesets that avoid ordering scenarios - but I can definitely see people requiring this for their own html
in the future.



git-svn-id: http://repo/sysadmin/perlmodules/CSS-Inliner@2640 49aadcc1-c794-da11-82e1-00902723a867